* compilling fix
git-svn-id: svn://darghos.com.br/darghos/tfs/8.6@2751 5021ead5-42e4-5646-a3ec-22fc1710fdc7

+++ darghos_pvp.cpp
@@ -474,7 +474,7 @@ BattlegrondRetValue Battleground::kickPlayer(Player* player, bool force)
 		delete playerInfo.statistics;
 
 		team->players.erase(player->getID());
-		team->levelSum = std::max(team->levelSum - player->getLevel(), 0);
+		team->levelSum = std::max((int32_t)(team->levelSum - player->getLevel()), 0);
 	}
 	else 
 	{
